requirements.txt
requests==2.32.3
python-dotenv==1.0.1
openpyxl==3.1.5